Fix bug with timecounter integer division when printing count result (should be floating point division)

+++ src/main/utility/TimeCounter.java
@@ -105,7 +105,7 @@ public class TimeCounter {
     }
 
     private static double per(int value) {
-        return value / iterations;
+        return (double)value / iterations;
     }
 
     public static void print() {
